+++ lib/bindings/auto-detect.js
@@ -1,6 +1,6 @@
 'use strict';
 
-const debug = require('debug')('serialport:binding');
+const debug = require('debug')('serialport:binding:auto-detect');
 
 /**
  * @module serialport
@@ -21,15 +21,15 @@ const debug = require('debug')('serialport:binding');
 switch (process.platform) {
   case 'win32':
     debug('loading WindowsBinding');
-    module.exports = require('./bindings-win32');
+    module.exports = require('./win32');
     break;
   case 'darwin':
     debug('loading DarwinBinding');
-    module.exports = require('./bindings-darwin');
+    module.exports = require('./darwin');
     break;
   default:
     debug('loading LinuxBinding');
-    module.exports = require('./bindings-linux');
+    module.exports = require('./linux');
 }
 
 /**
