+++ lib/bindings/darwin.js
@@ -27,7 +27,7 @@ function DarwinBinding(opt) {
 
 DarwinBinding.list = binding.list;
 
-DarwinBinding.prototype.read = require('./read-unix');
+DarwinBinding.prototype.read = require('./unix-read');
 
 DarwinBinding.prototype.open = function(path, options, cb) {
   if (!path) {
