+++ lib/bindings/linux-list.js
