+++ lib/bindings/linux.js
@@ -25,9 +25,9 @@ function LinuxBinding(opt) {
   this.fd = null;
 };
 
-LinuxBinding.list = require('./list-linux');
+LinuxBinding.list = require('./linux-list');
 
-LinuxBinding.prototype.read = require('./read-unix');
+LinuxBinding.prototype.read = require('./unix-read');
 
 LinuxBinding.prototype.open = function(path, options, cb) {
   if (!path) {
