+++ lib/bindings/mock.js
