+++ lib/bindings/win32.js
