+++ lib/index.js
@@ -6,7 +6,7 @@
  */
 
 const SerialPort = require('./serialport');
-const Binding = require('./bindings-auto-detect');
+const Binding = require('./bindings/auto-detect');
 const parsers = require('./parsers');
 
 SerialPort.Binding = Binding;
