+++ lib/parsers/index.js
@@ -53,7 +53,7 @@ parser.on('data', console.log);
  */
 
 module.exports = {
-  Readline: require('./parser-readline'),
-  Delimiter: require('./parser-delimiter'),
-  ByteLength: require('./parser-byte-length')
+  Readline: require('./readline'),
+  Delimiter: require('./delimiter'),
+  ByteLength: require('./byte-length')
 };
