+++ lib/parsers/readline.js
@@ -1,5 +1,5 @@
 'use strict';
-const DelimiterParser = require('./parser-delimiter');
+const DelimiterParser = require('./delimiter');
 const inherits = require('util').inherits;
 
 function ReadlineParser(options) {
