+++ test/manual-testing/memory-binding.js
@@ -6,7 +6,7 @@ if (!port) {
   process.exit(1);
 }
 
-// var Binding = require('../../lib/bindings-mock');
+// var Binding = require('../../lib/bindings/mock');
 // Binding.createPort(port);
 const Binding = require('../../').Binding;
