+++ test/manual-testing/performance.js
@@ -8,7 +8,7 @@ if (!port) {
   process.exit(1);
 }
 
-// var Binding = require('../../lib/bindings-mock');
+// var Binding = require('../../lib/bindings/mock');
 // Binding.createPort(port);
 // SerialPort.Binding = Binding;
 // debugger;
